Implement conditional navigation
Add a new option to config.yml ("conditional-navigation") to allow to
turn on conditional navigation.

In this mode, a goto entry in a branch of the main tree can be set to
a map.  The structure of this map is like so:

- conditional-score-outputs: true
- my-score-variable:
    value: 1
    node: "target-node"
- default: "default-target-node"

When goto is a map, "conditional-score-outputs" is optional, custom
variables are optional, and "default" is only required when there is
no "conditional-score-outputs" or custom variable.